feat: code org + added new fields to invoices
